custom_rnn_backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.31.0",
    "bcrypt==4.0.1",
    "fastapi>=0.122.0",
    "loguru>=0.7.3",
    "numpy>=2.3.5",
    "pandas>=2.3.3",
    "passlib[bcrypt]>=1.7.4",
    "psycopg>=3.2.13",
    "psycopg-binary>=3.2.13",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.20",
    "scikit-learn>=1.7.2",
    "sqlalchemy>=2.0.44",
    "torch>=2.9.1",
    "uvicorn>=0.38.0",
]

[build-system]
requires = ["uv_build>=0.8.10,<0.9.0"]
build-backend = "uv_build"

[project.scripts]
drop-db = "src.db.initial:drop_tables"
create-db = "src.db.initial:create_tables"

[tool.uv.build-backend]
module-name = "src"
module-root = ""



